feat(variable_load): Add code to import and print variable 'a'
Added code to the  5-variable_load.py  file to import the variable 'a' from the  variable_load_5  module and print its value. The code follows the provided requirements, including not using  *  for importing or  __import__ . The script is designed to be executed directly and will print the value of 'a' as 98.