feat(print_list_integer): add function to print integers in a list
Add the `print_list_integer` function to the Python script. This function takes a list of integers as input and prints each integer on a separate line using `str.format()`. The function is implemented using a for loop to iterate over the elements of the list. The function also includes a default parameter for the list. This commit improves the functionality of the script and makes it easier to print integers in a list.