feat(no_c): add function to remove 'c' and 'C' characters from string
Add the `no_c` function to the Python script. This function takes a string as input and removes all occurrences of the characters 'c' and 'C' from the string. The function creates a translation dictionary `my_dict` that maps the Unicode values of the characters 'c' and 'C' to None. Then, it uses the `translate()` method of the string to remove the characters 'c' and 'C' based on the translation dictionary. This commit enhances the functionality of the script by providing an efficient way to remove specific characters from a string.