feat(divisible_by_2): add function to check divisibility by 2 in a list
Add the `divisible_by_2` function to the Python script. This function takes a list of integers as input and uses the `map()` function to apply a lambda function to each element of the list. The lambda function checks if each element is divisible by 2 and returns `True` if it is, otherwise it returns `False`. The `map()` function applies the lambda function to each element and returns a map object, which is then converted into a list using the `list()` function. Finally, the function returns the resulting list of divisibility by 2. This commit enhances the functionality of the script by providing an efficient way to check divisibility by 2 in a list.