feat(switch): add code to switch values of variables
Add code to the Python script that switches the values of variables `a` and `b`. The initial values of `a` and `b` are provided, and the code uses tuple packing and unpacking to efficiently swap their values. After the switch, the updated values of `a` and `b` are printed using `str.format()`. This commit enhances the functionality of the script by allowing efficient switching of variable values.