feat(function): Add safe_print_integer_err function
Add safe_print_integer_err function to handle printing of an integer value. The function efficiently handles different types of input values, such as integers or strings, and prints the integer value followed by a new line. It returns True if the value is an integer and has been printed correctly. If the value is not an integer, it returns False and prints an error message to stderr, preceded by 'Exception:'. The function uses try and except blocks to handle exceptions and utilizes '{:d}'.format() to print the integer value. The sys module is imported to access the sys.exc_info() function, which provides information about the current exception being handled.