feat(class): Add Square class
Add Square class to represent a square. The Square class efficiently includes a private instance attribute called size. It defines getter and setter methods for the size attribute to retrieve and set the value with type and value validations. The class also includes a public instance method called area to calculate the current area of the square. No external modules are imported.