feat(class): Add Node and SinglyLinkedList classes
Add Node and SinglyLinkedList classes to implement a singly linked list. The Node class represents a node in the linked list and has private instance attributes data and next_node, along with their respective getters and setters. The SinglyLinkedList class represents the linked list itself and has a private instance attribute head. It also includes methods to insert a new node into the correct sorted position and provide a string representation of the linked list.