feat(class): Add Rectangle class
Add Rectangle class to represent a rectangle. The class efficiently handles attributes such as number_of_instances and print_symbol. It initializes the class attributes and provides methods for calculating area and perimeter, as well as string representation of the rectangle. The class also includes a destructor method to print a message upon deletion of a Rectangle instance.