docs(project): Add Alx Higher Level Programming - Python Everything is Object project
Add the Alx Higher Level Programming - Python Everything is Object project to the repository. The project consists of a series of tasks, each contained in a separate file within the  `0x09-python-everything_is_object`  directory. Each file corresponds to a specific task and contains answers to questions related to Python object behavior. The project overview, task descriptions, contributors, and acknowledgments are included in the documentation.