docs(file): Add answer to task 14
Add answer to task 14 in the file  `14-answer.txt` . The script will print  `[1, 2, 3, 4]`  because  `l1`  and  `l2`  are referencing the same list object. When  `l1.append(4)`  is called, it modifies the list in place, and since  `l2`  is also referencing the same list object, the change is reflected in  `l2`  as well.