fix(answer): Update answer for 26-empty_is_not_empty
Update the answer for the script in 26-empty_is_not_empty to clarify that it prints 'True'. The script creates two empty tuples  `a`  and  `b` , and then checks if  `a`  is the same object as  `b`  using the  `is`  operator. Since both tuples are empty and immutable objects, they refer to the same object in memory, resulting in the output 'True'.