fix(answer): Update answer for 27-still_the_same
Update the answer for the script in 27-still_the_same to clarify that the last line will not print 39926795932424'. The script assigns the result of concatenating  `a`  and  `[5]`  back to  `a` , which creates a new list object. Therefore, the  `id(a)`  will be different from the previous value. The last line will print a different ID.