fix(answer): Update answer for 28-same_or_not
Update the answer for the script in 28-same_or_not to clarify that the last line will print 139926795932424. The script uses the  `+=`  operator to append  `[4]`  to the list  `a`  in-place, modifying the existing list object. Therefore, the  `id(a)`  will remain the same. The last line will print the same ID.