feat(class): Add Rectangle class
Add Rectangle class that represents a rectangle and inherits from the BaseGeometry class. The class includes an initialization method to set the width and height of the rectangle. It also has an area method to calculate and return the area of the rectangle. The __str__ method returns a string representation of the rectangle in the format '[Rectangle] <width>/<height>'. The class is efficiently implemented with proper docstring comments throughout the code.