feat(function): Add add_attribute function
Add add_attribute function to add a new attribute to an object if possible. The function efficiently checks if the object has the __dict__ attribute, indicating if new attributes can be added. If the object does not have the __dict__ attribute, a TypeError exception is raised. The function uses setattr to set the attribute with the given name and value. The function is well-documented with a clear description, parameter explanations, and exception information.