feat(function): Add write_file function
Add write_file function to write a string to a text file (UTF8) and return the number of characters written. The function efficiently uses the  `open()`  function with the specified filename and encoding to open the file in write mode. It then uses the  `write()`  method of the file object to write the text to the file. The function returns the number of characters written. The docstring provides clear information about the purpose, arguments, and return value of the function.